Completed User API (refactor)
